pyproject.toml
[tool.poetry]
name = "test-django"
version = "0.1.0"
description = ""
authors = ["DmitriyDecadenz <dima.hubul@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
djangorestframework = "^3.15.1"
aiogram = "^3.7.0"
django-environ = "^0.11.2"
psycopg2 = "^2.9.9"
drf-spectacular = "^0.27.2"
django = "^5.0.6"
djoser = "^2.2.2"
httpx = "^0.27.0"
django-filter = "^24.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
